Separated show-counter tests into its own file
