+++ tests/like-counter.test.js
@@ -1,26 +1,15 @@
 /**
  * @jest-environment jsdom
  */
-import { getLikesCount, showsCount } from '../src/modules/counters';
-
-describe('Gets correct count of items based on a transformed array', () => {
-  test('showCount equal to four based on showlist', () => {
-    const showlist = ['show1', 'show2', 'show3', 'show4'];
-    expect(showsCount(showlist)).toBe(4);
-  });
-  test('showCount not equal to 4 based on showlist', () => {
-    const showlist = ['show1', 'show2', 'show3', 'show4'];
-    showlist.push('show5');
-    expect(showsCount(showlist)).toBe(5);
-  });
-});
+import { getLikesCount } from '../src/modules/counters';
 
 describe('Gets correct count of likes based on the item', () => {
   test('likeCount equals to 50', () => {
     document.body.innerHTML = '<div><ul class="container"></li></div>';
     const likesContainer = document.querySelector('.container');
     const likeBtn = document.createElement('div');
-    likeBtn.innerHTML = '<button id="1" class= "like-btn"><i class="fas fa-heart"></i></button>';
+    likeBtn.innerHTML =
+      '<button id="1" class= "like-btn"><i class="fas fa-heart"></i></button>';
     likeBtn.id = '1';
     const item1 = { item_id: '1', likes: 50, id: 1 };
     const item2 = { item_id: '2', likes: 500, id: 2 };
@@ -34,7 +23,8 @@ describe('Gets correct count of likes based on the item', () => {
     document.body.innerHTML = '<div><ul class="container"></li></div>';
     const likesContainer = document.querySelector('.container');
     const likeBtn = document.createElement('div');
-    likeBtn.innerHTML = '<button id="1" class= "like-btn"><i class="fas fa-heart"></i></button>';
+    likeBtn.innerHTML =
+      '<button id="1" class= "like-btn"><i class="fas fa-heart"></i></button>';
     likeBtn.id = '2';
     const item1 = { item_id: '1', likes: 50, id: 1 };
     const item2 = { item_id: '2', likes: 500, id: 2 };
@@ -48,7 +38,8 @@ describe('Gets correct count of likes based on the item', () => {
     document.body.innerHTML = '<div><ul class="container"></li></div>';
     const likesContainer = document.querySelector('.container');
     const likeBtn = document.createElement('div');
-    likeBtn.innerHTML = '<button id="1" class= "like-btn"><i class="fas fa-heart"></i></button>';
+    likeBtn.innerHTML =
+      '<button id="1" class= "like-btn"><i class="fas fa-heart"></i></button>';
     likeBtn.id = '3';
     const item1 = { item_id: '1', likes: 500, id: 1 };
     const item2 = { item_id: '2', likes: 500, id: 2 };
